[FEAT] 채팅 저장 기능 중간커밋

+++ naegamaja/src/main/java/naegamaja_server/naegamaja/domain/room/controller/RoomRestController.java
@@ -25,7 +25,7 @@ public class RoomRestController {
     }
 
     @PostMapping("/create")
-    public int createRoom(@RequestHeader String authorization, @RequestBody RoomRequest.CreateRoomRequest request) {
+    public Long createRoom(@RequestHeader String authorization, @RequestBody RoomRequest.CreateRoomRequest request) {
         System.out.println("request = " + request);
         return roomService.createRoom(request, authorization);
     }
